refactor: Update environment configuration and created files for helper functions previously inside `extract_procedure`
