+++ .env.example
@@ -15,13 +15,13 @@ DB_PORT=5432             # Port number for PostgreSQL server (default: 5432)
 GEMINI_API_KEY=          # Required: Your Google AI Studio API key
 
 # Model selection and parameters
-# MAIN_MODEL=gemini-2.0-pro-exp-02-05      # Optional: Gemini model to use (defaults to gemini-2.0-flash if not specified)
+MAIN_MODEL=gemini-2.5-pro-exp-03-25      # Optional: Gemini model to use (defaults to gemini-2.0-flash if not specified)
 
 # MAIN_MODEL_TEMPERATURE=                  # Optional: Controls randomness in output (defaults to 0)
                                            # Range: 0.0 (deterministic) to 1.0 (creative)
 
 
-# ALTERNATIVE_MODEL=                       # Optional: Alternative model for comparison (defaults to None)
+ALTERNATIVE_MODEL=gemini-2.0-flash                       # Optional: Alternative model for comparison (defaults to MAIN_MODEL)
 
 # ALTERNATIVE_MODEL_TEMPERATURE=           # Optional: Controls randomness in output (defaults to 0)
                                            # Range: 0.0 (deterministic) to 1.0 (creative)
